refactor(dg): add sortStore
