+++ packages/data-grid/src/store.ts
@@ -1,13 +1,12 @@
 import { createComputed, createSignal } from '@cerberus-design/signals'
 import { DEFAULT_THEME } from './const'
-import { createDataStore, createFilterStore, createPaginationStore } from './stores'
-import type {
-  GridOptions,
-  GridStore,
-  InternalColumn,
-  SortState,
-  ThemeOptions,
-} from './types'
+import {
+  createDataStore,
+  createFilterStore,
+  createPaginationStore,
+  createSortStore,
+} from './stores'
+import type { GridOptions, GridStore, InternalColumn, ThemeOptions } from './types'
 import { determineInitialCount } from './utils'
 
 /**
@@ -21,52 +20,15 @@ export function createGridStore<TData>(options: GridOptions<TData>): GridStore<T
   // 2. Feature Stores
   const paginationStore = createPaginationStore(options)
   const filterStore = createFilterStore(dataStore)
+  const sortStore = createSortStore({
+    columns: dataStore.columns,
+    filteredRows: filterStore.filteredRows,
+  })
 
   const [containerWidth, setContainerWidth] = createSignal<number>(0)
 
   const [pending, setPending] = createSignal<boolean>(options.pending ?? false)
   const [hasSkeleton] = createSignal<boolean>(options.overlays?.pending === 'skeleton')
-
-  const [sorting, setSorting] = createSignal<SortState[]>([])
-
-  const sortedRows = createComputed(() => {
-    const filteredRows = filterStore.filteredRows()
-    const currentRows = [...filteredRows]
-    const sortState = sorting()
-    const cols = dataStore.columns()
-
-    if (sortState.length === 0) return currentRows
-
-    return currentRows.sort((a, b) => {
-      for (const sort of sortState) {
-        const col = cols.find((c) => c.id === sort.id)
-        if (!col) continue
-
-        const valA = col.getValue(a) as TData[keyof TData]
-        const valB = col.getValue(b) as TData[keyof TData]
-
-        if (valA === valB) continue // Move to next tie-breaker if equal
-
-        // Use custom comparator if provided
-        let comparison = 0
-        const customComparator =
-          typeof col.original.features?.sort === 'object'
-            ? col.original.features.sort.comparator
-            : undefined
-
-        if (customComparator) {
-          comparison = customComparator(valA, valB)
-        } else {
-          // Fallback: Default JS Comparison
-          comparison = valA > valB ? 1 : -1
-        }
-
-        // Invert the result if we are sorting descending
-        return sort.desc ? -comparison : comparison
-      }
-      return 0
-    })
-  })
 
   // Derived pagination - Ark handles the rest
   const rowCount = createComputed(() => {
@@ -81,7 +43,7 @@ export function createGridStore<TData>(options: GridOptions<TData>): GridStore<T
 
   const visibleRows = createComputed(() => {
     const size = paginationStore.pageSize()
-    const rows = sortedRows()
+    const rows = sortStore.sortedRows()
 
     if (paginationStore.isServerPaginated()) return rows
 
@@ -190,9 +152,9 @@ export function createGridStore<TData>(options: GridOptions<TData>): GridStore<T
     ...dataStore,
     ...paginationStore,
     ...filterStore,
+    ...sortStore,
     rowCount,
     visibleRows,
-    sorting,
     pending,
     hasSkeleton,
     pageCount,
@@ -205,52 +167,9 @@ export function createGridStore<TData>(options: GridOptions<TData>): GridStore<T
       setPending(newState)
     },
 
-    setSort: (colId, direction, multi = false) => {
-      if (direction === null) {
-        setSorting(sorting().filter((s) => s.id !== colId))
-        return
-      }
-
-      const current = sorting()
-      const newSort = { id: colId, desc: direction === 'desc' }
-
-      if (multi) {
-        const existingIndex = current.findIndex((s) => s.id === colId)
-        if (existingIndex >= 0) {
-          const next = [...current]
-          next[existingIndex] = newSort
-          setSorting(next)
-        } else {
-          setSorting([...current, newSort])
-        }
-      } else {
-        // Single sort clears all other sorts
-        setSorting([newSort])
-      }
-    },
-
     togglePinned: (colId, state) => {
       const col = dataStore.columns().find((c) => c.id === colId)
       if (col) col.setPinned(state ?? false)
-    },
-
-    toggleSort: (colId, multi) => {
-      const current = sorting()
-      const exists = current.findIndex((s) => s.id === colId) !== -1
-
-      const updatedSort = current.map((s) => {
-        if (s.id === colId) {
-          return { ...s, desc: !s.desc }
-        }
-        return s
-      })
-
-      if (exists) {
-        setSorting(multi ? [...current, ...updatedSort] : [...updatedSort])
-      } else {
-        const newSort = { id: colId, desc: true }
-        setSorting(multi ? [...current, newSort] : [newSort])
-      }
     },
 
     setContainerWidth: (w: number) => {
