+++ packages/data-grid/src/stores/index.ts
@@ -1,3 +1,4 @@
 export * from './data'
 export * from './pagination'
 export * from './filter'
+export * from './sort'
